requirements.txt
asgiref==3.9.1
asttokens==3.0.0
black==25.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.9.2
cryptography==45.0.5
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.3
dnspython==2.7.0
drf-yasg==1.21.10
eventlet==0.40.1
executing==2.2.0
flake8==7.3.0
gevent==25.5.1
greenlet==3.2.3
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
setuptools==80.9.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
zope.event==5.1.1
zope.interface==7.2
